+++ njnu-classroom-core/src/main/kotlin/cn/repigeons/njnu/classroom/model/dto/EmptyClassroomFeedbackDTO.kt
@@ -9,7 +9,7 @@ data class EmptyClassroomFeedbackDTO(
     @Schema(description = "星期")
     val weekday: Weekday,
     @Schema(description = "节次")
-    val jc: Short,
+    val jc: Int,
     @Schema(description = "教学楼名称")
     val jxlmc: String,
     @Schema(description = "查询结果集")
