+++ njnu-classroom-core/src/main/kotlin/cn/repigeons/njnu/classroom/service/EmptyClassroomService.kt
@@ -5,11 +5,11 @@ import cn.repigeons.njnu.classroom.model.vo.EmptyClassroomVO
 import java.util.concurrent.Future
 
 interface EmptyClassroomService {
-    suspend fun getEmptyClassrooms(jxlmc: String, weekday: Weekday, jc: Short): List<EmptyClassroomVO>
+    suspend fun getEmptyClassrooms(jxlmc: String, weekday: Weekday, jc: Int): List<EmptyClassroomVO>
     fun feedback(
         jxlmc: String,
         weekday: Weekday,
-        jc: Short,
+        jc: Int,
         results: List<EmptyClassroomVO>,
         index: Int,
     ): Future<*>
